Changing wording of project global variable override.
On reading this page as part of making this change, I see that it links
to the project config docs where project.globals are mentioned and that
links to further documentation on project globals that fully explains
it.  But I looked at this top level a couple times and missed this
wording and the link.  I'm thinking this wording might have helped,
but perhaps this is more my problem than a docs problem.  :-)