Add warning about comments for jaas-ldap.conf
When reconfiguring our rundeck ldap integration I restarted and started receiving 503 errors after login.

The root cause of this was found to be a comment in the jaas-ldap.conf.  When searching for more information we found documentation of this in the confluent docs: https://docs.confluent.io/5.4.5/control-center/security/c3-auth-ldap.html