Update gcp.md
I went to look for the docs on the "capturing snapshot" gcp plugin because I was struggling to get it working and didn't know what to do, and noticed that it and a few other gcp plugins were missing, so I thought I'd make a PR to add them. I have since figured out that the plugin was failing because I entered a space in the name of the snapshot which I didn't know wasn't allowed, so i also added the detail that you shouldn't use spaces.