Update resource-yaml-v13.md
Updates the documentation for the YAML resource definition to remove `sshport` as a known optional entry, as it does not actually have an effect. It appears that one has to provide the port by appending `:<port>` to the `hostname` value, as noted for the `hostname` attribute. 

I notices that `sshport` is not documented within the XML or JSON definition for a node resource either, making me think the fact that `sshport` has no effect was not simply just a bug.